Decouple revocation_notification from Issuer handler
